requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic>=2.0
openai==1.30.1
python-dotenv==1.0.1
httpx==0.27.0
